+++ md5/string_operations.c
@@ -43,7 +43,7 @@ char	*prepare_input_string(const char *str, uint64_t len)
 	int			new_len;
 	uint64_t	bits_len;
 
-	r = len;
+	r = (int)len;
 	new_len = ((((r + 8) / 64) + 1) * 64) - 8;
 	s = malloc(new_len + 8);
 	if (s == NULL)
