+++ utils/ft_strjoin.c
@@ -23,7 +23,7 @@ char	*ft_strjoin(const char *s1, const char *s2, int n, int free_first)
 	len1 = 0;
 	len2 = n;
 	if (s1 != NULL)
-		len1 = ft_strlen(s1);
+		len1 = (int)ft_strlen(s1);
 	len = len1 + len2;
 	res = malloc(sizeof(char) *(len + 1));
 	if (res == NULL)
